+++ src/main/java/com/github/dockerjava/api/command/InspectContainerResponse.java
@@ -20,6 +20,7 @@ import com.github.dockerjava.api.model.VolumeBinds;
 import com.github.dockerjava.api.model.VolumeRW;
 import com.github.dockerjava.api.model.VolumesRW;
 import com.github.dockerjava.core.RemoteApiVersion;
+import org.apache.commons.lang.builder.ToStringStyle;
 
 /**
  *
@@ -52,6 +53,12 @@ public class InspectContainerResponse {
 
     @JsonProperty("HostsPath")
     private String hostsPath;
+
+    /**
+     * @since {@link RemoteApiVersion#VERSION_1_17}
+     */
+    @JsonProperty("LogPath")
+    private String logPath;
 
     @JsonProperty("Id")
     private String id;
@@ -106,7 +113,7 @@ public class InspectContainerResponse {
     }
 
     public Integer getSizeRootFs() {
-        return  sizeRootFs;
+        return sizeRootFs;
     }
 
     public String getCreated() {
@@ -168,6 +175,11 @@ public class InspectContainerResponse {
         return hostsPath;
     }
 
+    @CheckForNull
+    public String getLogPath() {
+        return logPath;
+    }
+
     public String getName() {
         return name;
     }
@@ -206,7 +218,7 @@ public class InspectContainerResponse {
 
     @Override
     public String toString() {
-        return ToStringBuilder.reflectionToString(this);
+        return ToStringBuilder.reflectionToString(this, ToStringStyle.SHORT_PREFIX_STYLE);
     }
 
     @JsonIgnoreProperties(ignoreUnknown = true)
@@ -290,6 +302,13 @@ public class InspectContainerResponse {
         @CheckForNull
         @JsonProperty("FinishedAt")
         private String finishedAt;
+
+
+        /**
+         * @since Docker version 1.12
+         */
+        @JsonProperty("Health")
+        private HealthState health;
 
         /**
          * See {@link #status}
@@ -377,6 +396,10 @@ public class InspectContainerResponse {
         @CheckForNull
         public String getFinishedAt() {
             return finishedAt;
+        }
+
+        public HealthState getHealth() {
+            return health;
         }
 
         @Override