+++ src/main/java/com/github/dockerjava/netty/InvocationBuilder.java
@@ -27,6 +27,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.CountDownLatch;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -66,6 +67,63 @@ public class InvocationBuilder {
         @Override
         public void onNext(T object) {
             result = object;
+        }
+    }
+
+    public class SkipResultCallback extends ResultCallbackTemplate<ResponseCallback<Void>, Void> {
+        @Override
+        public void onNext(Void object) {
+        }
+    }
+
+    /**
+     * Implementation of {@link ResultCallback} with the single result event expected.
+     */
+    public static class AsyncResultCallback<A_RES_T>
+            extends ResultCallbackTemplate<AsyncResultCallback<A_RES_T>, A_RES_T> {
+
+        private A_RES_T result = null;
+
+        private final CountDownLatch resultReady = new CountDownLatch(1);
+
+        @Override
+        public void onNext(A_RES_T object) {
+            onResult(object);
+        }
+
+        private void onResult(A_RES_T object) {
+            if (resultReady.getCount() == 0) {
+                throw new IllegalStateException("Result has already been set");
+            }
+
+            try {
+                result = object;
+            } finally {
+                resultReady.countDown();
+            }
+        }
+
+        @Override
+        public void close() throws IOException {
+            try {
+                super.close();
+            } finally {
+                resultReady.countDown();
+            }
+        }
+
+        /**
+         * Blocks until {@link ResultCallback#onNext(Object)} was called for the first time
+         */
+        @SuppressWarnings("unchecked")
+        public A_RES_T awaitResult() {
+            try {
+                resultReady.await();
+            } catch (InterruptedException e) {
+                throw new RuntimeException(e);
+            }
+            getFirstError();
+            return result;
         }
     }
 
@@ -197,7 +255,7 @@ public class InvocationBuilder {
 
         Channel channel = getChannel();
 
-        ResponseCallback<InputStream> callback = new ResponseCallback<InputStream>();
+        AsyncResultCallback<InputStream> callback = new AsyncResultCallback<>();
 
         HttpResponseHandler responseHandler = new HttpResponseHandler(requestProvider, callback);
         HttpResponseStreamHandler streamHandler = new HttpResponseStreamHandler(callback);
@@ -401,6 +459,31 @@ public class InvocationBuilder {
         channel.pipeline().addLast(new JsonObjectDecoder());
         channel.pipeline().addLast(jsonResponseHandler);
 
+        postChunkedStreamRequest(requestProvider, channel, body);
+    }
+
+    public void postStream(InputStream body) {
+        SkipResultCallback resultCallback = new SkipResultCallback();
+
+        HttpRequestProvider requestProvider = httpPostRequestProvider(null);
+
+        Channel channel = getChannel();
+
+        HttpResponseHandler responseHandler = new HttpResponseHandler(requestProvider, resultCallback);
+
+        channel.pipeline().addLast(new ChunkedWriteHandler());
+        channel.pipeline().addLast(responseHandler);
+
+        postChunkedStreamRequest(requestProvider, channel, body);
+
+        try {
+            resultCallback.awaitCompletion();
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private void postChunkedStreamRequest(HttpRequestProvider requestProvider, Channel channel, InputStream body) {
         HttpRequest request = requestProvider.getHttpRequest(resource);
 
         // don't accept FullHttpRequest here
@@ -423,7 +506,7 @@ public class InvocationBuilder {
 
         Channel channel = getChannel();
 
-        ResponseCallback<InputStream> resultCallback = new ResponseCallback<InputStream>();
+        AsyncResultCallback<InputStream> resultCallback = new AsyncResultCallback<>();
 
         HttpResponseHandler responseHandler = new HttpResponseHandler(requestProvider, resultCallback);