+++ src/test/java/com/github/dockerjava/netty/exec/AttachContainerCmdExecTest.java
@@ -6,13 +6,14 @@ import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.isEmptyString;
 import static org.hamcrest.Matchers.not;
 
-import java.io.ByteArrayInputStream;
 import java.io.File;
-import java.io.InputStream;
+import java.io.PipedInputStream;
+import java.io.PipedOutputStream;
 import java.lang.reflect.Method;
-import java.util.concurrent.TimeUnit;
+import java.util.Locale;
 
 import org.testng.ITestResult;
+import org.testng.SkipException;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeMethod;
@@ -31,6 +32,10 @@ public class AttachContainerCmdExecTest extends AbstractNettyDockerClientTest {
 
     @BeforeTest
     public void beforeTest() throws Exception {
+        // io.netty.channel.epollNative#loadNativeLibrary only support Linux
+        if (!System.getProperty("os.name").toLowerCase(Locale.UK).trim().startsWith("linux")) {
+            throw new SkipException("only support Linux");
+        }
         super.beforeTest();
     }
 
@@ -93,8 +98,6 @@ public class AttachContainerCmdExecTest extends AbstractNettyDockerClientTest {
 
         dockerClient.startContainerCmd(container.getId()).exec();
 
-        Thread.sleep(SECONDS.toMillis(10)); //wait bash initialisation
-
         InspectContainerResponse inspectContainerResponse = dockerClient.inspectContainerCmd(container.getId()).exec();
 
         assertTrue(inspectContainerResponse.getState().getRunning());
@@ -107,15 +110,20 @@ public class AttachContainerCmdExecTest extends AbstractNettyDockerClientTest {
             }
         };
 
-        InputStream stdin = new ByteArrayInputStream((snippet + "\n").getBytes());
+        PipedOutputStream out = new PipedOutputStream();
+        PipedInputStream in = new PipedInputStream(out);
 
         dockerClient.attachContainerCmd(container.getId())
                 .withStdErr(true)
                 .withStdOut(true)
                 .withFollowStream(true)
-                .withStdIn(stdin)
-                .exec(callback)
-                .awaitCompletion(15, SECONDS);
+                .withStdIn(in)
+                .exec(callback);
+
+        out.write((snippet + "\n").getBytes());
+        out.flush();
+
+        callback.awaitCompletion(15, SECONDS);
         callback.close();
 
         assertThat(callback.toString(), containsString(snippet));