+++ CHANGELOG.md
@@ -1,9 +1,47 @@
 Change Log
 ===
-## Next
+## (dev) 3.1.0 (reserverd for swarm features)
+ 
+## 3.0.11
+- Add labels and attachable properties to network.
+- Set default socket timeout for RequestConfig.
+- Netty skip instead of throw error on non-linux os.
+- Clean tmp file after upload.
+- Filters ignore application/x-tar.
+- Allow user to call connectionManager's closeIdleConnections.
+ 
+## 3.0.10
+- Support for cache-from in build image command
+- Allow multiple tags in build image command 
+- Custom `db` logging type 
+- Allow an explicit Dockerfile location string to be specified to theuild command
+- Fix image build for docker 17 with 'tagged' word.
+ 
+## 3.0.9
+- NettyDockerCmdExecFactory ignores API version 
+- exclude commons-logging from httpclient since docker-java uses slf4j/logback
+- Generate OSGi compliant manifest
+- AuthResponse may contains token.
+
+## 3.0.8
+ - Use TLSv1.2 by default
+ - Health api
+ - Labels
+ - Support for multiple certificates 
+
+## 3.0.7
+ * https://github.com/docker-java/docker-java/milestone/17?closed=1
+ * HostConfig pidLimits
+ * Label image during build
+ * Expose 'User' property on ExecCreateCmd #707 #708
 
 ## 3.0.6
- * Make all models Serializable
+ * Fixed issue with jersey and unix domain sockets.
+ * [#703](https://github.com/docker-java/docker-java/pull/703) Allow to configure connection pool timeout.
+ * Make all models Serializable.
+ * [NETTY] Fix loadImage responce on 1.24 API.
+ * LogPath field for inspect container.
+ * [#700] (https://github.com/docker-java/docker-java/pull/700) Bugfix:donot throw RuntimeException when a error occured in awaitCompletion(long,TimeUnit)
 
 ## 3.0.5
  * Events updated to 1.24 API model.